use ProfitCalculations.cs

+++ Client/Shared/Shared.Models/PropertyPredictionResult.cs
@@ -8,27 +8,27 @@ namespace CapitalClue.Frontend.Shared.Models;
 
 public class PropertyPredictionResult
 {
-    public float Gain { get; set; }
-    public float ConfidenceLowerGain { get; set; }
-    public float ConfidenceUpperGain { get; set; }
+    public double Gain { get; set; }
+    public double ConfidenceLowerGain { get; set; }
+    public double ConfidenceUpperGain { get; set; }
 
-    public Dictionary<int, float> ForeCastIndex { get; set; }
-    public Dictionary<int, float> ConfidenceLowerBound { get; set; }
-    public Dictionary<int, float> ConfidenceUpperBound { get; set; }
+    public Dictionary<int, double> ForeCastIndex { get; set; }
+    public Dictionary<int, double> ConfidenceLowerBound { get; set; }
+    public Dictionary<int, double> ConfidenceUpperBound { get; set; }
 
     public PropertyPredictionResult()
     {
-        ForeCastIndex = new Dictionary<int, float>();
-        ConfidenceLowerBound = new Dictionary<int, float>();
-        ConfidenceUpperBound = new Dictionary<int, float>();
+        ForeCastIndex = new Dictionary<int, double>();
+        ConfidenceLowerBound = new Dictionary<int, double>();
+        ConfidenceUpperBound = new Dictionary<int, double>();
     }
 }
 
 public class PropertyPurchaseInfo
 {
-    public float InterestRate { get; set; }
-    public float PropertyPrice { get; set; }
-    public float PropertyMonthlyPayment { get; set; }
-    public float DownPayment { get; set; }
+    public double InterestRate { get; set; }
+    public double PropertyPrice { get; set; }
+    public double PropertyMonthlyPayment { get; set; }
+    public double DownPayment { get; set; }
     public int MortgageTerm { get; set; }
 }
